Factor Fields->conversion code into reusable method
